+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [firstDigit](docs/firstDigit.md)
 - [pow10](docs/pow10.md)
+- .round and .multiply to [Point](docs/Point.md)
 
 ## [1.1.0] - 2020-01-31
 ### Added